backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
python-dotenv
# tavily-python
langchain-tavily